refactor: move minicalc tests under their own package
